requirements.txt
docspec
docspec-python
pytest
pyyaml